host: improve wrapping command with sudo
"cmd" is now a shell script. Just prepending "sudo" is not correct.
Instead, we need to go through `sudo sh -c`.

Note that this breaks allowing only certain commands in /etc/sudoers.
Sudo now always sees "sh -c ..." and filtering that seems hard.
In general, it's undefined which commands CDA will want to (rightly)
call, so sudo only allowing certain commands does not work anyway. And
it's (no longer) supported.

The benefit is that this works with "cwd" and "env_extra" arguments and
with "cmd" being a shell script.