host: don't split command by newlines in Host._run_remote()
Host.run() has two implementations, _run_local() and _run_remote(). We
want them to behave as similar as possible.

Note that _run_remote() uses SSH, which can only pass a "-c '{cmd}'" to
the user's login shell. Also, various callers rely on this being a shell
script, for example we see

    ret = rh.run(f"rpm -qa | grep {package}")
    ...
    cmd = f"curl -k {e} --create-dirs > {wd}/{fn}"
    rh.run(cmd)

On the other hand, _run_local() interprets the command without a shell.

This will be unified next, so that the "cmd" argument are interpreted
the same way.

We will choose the semantics of "cmd" being the argument passed to the
shell, and not a string parsed by shlex.split() and executed directly
(execvpe()). That choice is made, because having a full blown shell is
more powerful (for user who want it) and the performance overhead is not
a concern for CDA (and for remote calls, the shell is what we start with
anyway).

However, _run_remote() was not really accepting plain shell scripts. It
did additional mangling of newlines. That behavior was introduced by commit
178389664248 ('Added support for VMs in remote nodes (i.e. not
localhost)'). I suspect, this was a wrong attempt to make the "cmd"
string be a plain program name and arguments. In that case, it should
have used shlex.split(). Doing that would have made it at least
consistent with Host._run_local(). Although, clearly some callers expect
the command to be a shell script.

Anyway. "cmd" shall be the shell script passed to "sh -c". Drop this
mangling.

Try:

    $ REMOTE=127.0.0.1 ; python -c 'if 1:
        import host;
        h = host.RemoteHost("'"$REMOTE"'");
        h.ssh_connect("root", "");
        r = h.run("whoami\nhostname\necho \"user: $USER\nhi\"");
        print(f"result: r={r.returncode};\nout=>>>\n{r.out}<<<\nerr=>>>\n{r.err}<<<\n")
    '

Only one caller cared about this newline mangling. Adjust that one.