host: log better message from host.run() and add options to adjust logging
1) Previously, host.run() would log the results like

    logger.log(log_level, ret_val)

That gives no indication of the command that was run. It seems bad when
reading the log. Instead, print a message including the command.

2) Also, some callers do

    logger.info(rsh.run(cmd))

which also don't show the context of the command. Add an option
log_level_result so we can (only) increase the logging level of the
result. Now we can instead do

    rsh.run(cmd, log_level_result=logging.INFO)

3) also, add "log_level_warn", to bump the logging level only on failure
further. In the past, virshPool.sh had a wrapper

    def run_and_log(cmd):
        r = self._host.run(cmd)
        logger.debug(cmd)
        if r.returncode:
            logger.warn(f"Ran {cmd} and got error: {r}")

This can now be replaced by

    rsh.run(cmd, log_level_warn=logging.WARN)

4) also, you can also combine them to always print the result,
but with warning level on failure.

    rsh.run(cmd, log_level_result=logging.INFO, log_level_warn=logging.WARN)

5) finally, add an optional "log_prefix" argument that is prepended to the
logging messages from host.run(). Logging prefixes are important for
giving context to log messages and for grouping similar messages.