k8sClient: add helper to login as kubeadmin and get token
In some cases, we will need the token. For example, for the internal
container registry, we want to `podman login`. But for that we need a
token, and from the kubeconfig certificate we don't get one.

Instead, we have the kubeadmin user and it's password. Use that, to get
a token.

In this case, call `oc login` for the kubeadmin. The problem with that
is, that it stores the login session in "~/.kube/config", which requires
us to logout afterwards. Which require us to login again with the
kubeconfig certificate to fix it. I feel, there should be a simpler way
(more stateless) to create a login token and return it. If you find it,
be welcome.

Try:

    $ CLUSTER_NAME=...
    $ KUBECONFIG=... \
      python -c 'if 1:
        import k8sClient

        k = k8sClient.K8sClient()

        token = k.kubeadmin_login(cluster_name="'"$CLUSTER_NAME"'")
        r = k.oc("whoami", with_kubeconfig=False)
        print(f">>> kubeadmin token: {token} (whoami: {r.out.strip()})")

        k.kubeadmin_logout()
        r = k.oc("whoami")
        print(f">>> whoami: r={r.returncode}, out={r.out.strip()}")
      '