k8sClient: add K8sClient.clusterca_trust() helper
The certificate of the cluster is self signed. That is annoying. Add a
way to trust it.

It's also kinda important when using the internal container registry.
Without it, fetching from that registry will fail.

Of course, it's pretty dangerous to install a certificate authority in
the system. And attacker might use it to spoof TLS connections on the
host. We deem this acceptable, because CDA is used for testing and CI.
Those are not production systems.

  $ CLUSTER_NAME=...
  $ KUBECONFIG=... \
    python -c 'if 1:
        import k8sClient, host, sys

        k = k8sClient.K8sClient()

        cert = k.clusterca_get()
        print(f"clusterca:\n\n{cert}")

        rsh = host.RemoteHost(sys.argv[1])
        rsh.ssh_connect("core", "")

        k.clusterca_trust_on_host("'"$CLUSTER_NAME"'",
                                  cert,
                                  rsh,
                                  sudo=False)
    ' \
    192.168.122.4