host: avoid hanging remote process by starving stderr
We cannot just read one of the output file descriptors fully. Otherwise,
the other file descriptor can fill up, blocking the process and hanging
the call.

We need to read them in parallel.

Add a loop. This is not great, because it polls and sleeps. paramiko
doesn't seem to make this easy. Maybe we should use the channels'
fileno() method to poll on. For now, this is good enough (probably).