clusterDeployer: trust the CA certificate of the cluster on all nodes
It is useful that the nodes in our cluster trust the CA certificate of
the cluster. Usually we can workaround that with `curl -k`, `oc login
--insecure-skip-tls-verify=true`, etc. However, in some cases that is
cumbersome. For example, when we push an image to the internal container
registry, then worker nodes will not be able to pull them due to TLS
validation error. Fix that by trusting adding the cluster's CA as
trusted.

Note that trusting a CA systemwide is potentially dangerous. An attacker
may be able to get the certificate and spoof any TLS connection.

In practice, we use this for testing and CI. This is not a concern on
our system.