clusterDeployer,clusterHost: avoid spamming log output with messages about waiting
The logging output is important for understanding what happens. Having the
script fill up the terminal output with repeated lines like

  2024-04-19 02:50:56 INFO: Not all workers ready (try #14). 1 are not ready yet.
  2024-04-19 02:51:27 INFO: Not all workers ready (try #15). 1 are not ready yet.
  ...

is not useful. Only log a distinct message once, and repeated messages
at DEBUG level.

While at it, don't log the "try_count". That is not a useful unit, for
one, you would need to multiply the "try_count" with the sleep time (10
seconds) to know the actually elapsed time. But also, that does not give
you reliably the elapsed time, because the process might take some time
beside sleeping.