host: ensure internalization of hosts takes into account the requested type
The way how Host.__new__() intercepts object creation and possibly
returns a cached instance seem not good.

In any case, the class of the requested object must be considered,
otherwise, an earlier call to

  host.RemoteHost(hostname)

will cause a Host object to be stored for "hostname". A subsequent
request to "create" another type will give the cached one:

  >>> import host
  >>> h1 = host.RemoteHost("host")
  >>> h2 = host.HostWithBF2("host")
  >>> h1
  <host.Host object at 0x7fddd21540e0>
  >>> h2
  <host.Host object at 0x7fddd21540e0>

This is still ugly, __new__() should not be misused for caching. For now
just fix the bug.