clustersConfig: parse HostConfig with structparse helpers
Make HostConfig an immutable dataclass. Also, for parsing use
structparse helpers.

Previously, we passed the (what we blindly presume to be a) dictionary
to the HostConfig constructor. Then do some error checking (but it was
easy to generate YAML so that HostConfig was set to bogus values, like
having the wrong type).
That could be tightened up with ktoolbox.common.dataclass_from_dict(),
which does more error checking and ensures that the dataclass is valid
according to it's type annotations. For input that was generated by a
tool, this would be the best way, because it's low effort and gives
strict(er) validation.

However, here we have user configuration. We want to
- accept variants of configuration (e.g. "network_api_port:\"auto\"" and
  "network_api_port:\"\"" to mean the same thing).
- have stricter validation then just checking the dataclass' type
  annotations.
- have better error messages when the configuration is wrong (including
  the YAML path to find which part is wrong).
- to get a normalized representation of the configuration, so that other
  layers of the code can depend on validation and normalization.
- be able to serialize the normalized representation back to YAML. This
  is useful for reviewing how the input gets normalized. Also, the
  normalized YAML is again valid configuration that can be parsed back to
  the same thing.

To get this, we need more effort (and lines of code). Implement this
with the structparse helpers.

For example, previously "pre_installed" was just a string. There was a
method is_preinstalled() which compared whether the string was literally
"true". That means, a value "yes" was silently treated as False. Any
bogus value also also treated as False. That is bad for user
configuration. Now, during parse we just parse a boolean, in various
forms of spelling and failing for bogus values. While such input
validation and normalization is also possible with
ktoolbox.common.dataclass_from_dict(), in more complex (and interesting
cases) it requires the effort that we put in here.

The benefit is that the parsing, validation and normalization is a
stateless early step (it's thus easy to review and test). When there is
a configuration error we fail early. Other layers of the code, can rely
on what we do.