clustersConfig: add "system_check()" method for validating ClusterConfig
Certain validations have side effects (like checking something on the
system). Those don't check the validity of the YAML, but rather further
how the parsed configuration validates on the system.

Rename the method "pre_check()" to "system_check()". That's the better
name.

Also, rename the "test_only" flag to "with_system_check", which defaults
to True. This is the way for the unit test to opt-out from running the
extra check.