clustersConfig: rework detection of external_port
ClustersConfig object is almost immutable, except for the
"external_port", which it will resolve in case it's unset/auto.

This is fine. So we have the immutable configuration (MainConfig,
ClustersConfig.cluster_config), while ClustersConfig also contains
some additional state (currently only the external_port).

Rework how the external port was detected. Instead of having a
ClustersConfig.prepare_external_port() that must be called before
accessing ClustersConfig.external_port, make it a getter
ClustersConfig.get_external_port().

If the getter is accessed the first time, it will resolve the auto port
automatically (or fail with an exception). On subsequent calls, it just
returns the previously validated value. From point of few of the caller,
this function returns always the same value (or fails).

Also, add a lock, since we initialized this property from potentially
multiple threads.