clustersConfig: validate node IP ranges while parsing the configuration
The cluster configuration MainConfig and it's nested objects are
immutable. For the most part, the properties just come from the YAML
configuration. In case of NodeConfig.real_ip_range, the value may
actually be injected via the "get_last_ip()" callback during
construction. So it does not only depend on the YAML, but it's still
provided early on, while parsing the YAML.

At that point, validating the IP addresses of the nodes is purely based
on the YAML itself (and the range potentially provided by get_last_ip()).
It's all known while parsing, and that is the moment when to validate
the IP addresses.

Also, fix the examples in the README, because the IP addresses are out of range.