clustersConfig: drop detection of last_ip for the IP range
The IP addresses are commonly configured in the cluster YAML. Also, the
"ip_range" can be configured there.

It seems odd, to fetch a range from virsh, and let that affect the
configuration.

Instead, the configuration YAML (and it's potential defaults) determine
the range. If virsh disagrees, we must reconfigure libvirt and not the
other way around.

Compare to "external_port: auto", where we also pick up external
configuration from the system. But for the IP range, we have the node IP
addresses in the YAML itself, so the detected range must match with what
we configured there. I think such configuration should not be
autodetected.

Also, because the very first time when you run CDA deploy
on a host, this will not pick up anything, because virsh is not
configured.

The benefit is, that the configuration only depends on the YAML file.
Always.