clustersConfig: move handling of secrets_path to ClustersConfig
ClustersConfig is mainly about parsing the YAML file. However, it's
convenient that it also holds related configuration options. It thus has
a "secrets_path" argument, that (in CDA) comes from the command line.

But the secrets path is only required with certain configurations. For
example, cluster-kind "iso" does not require it (unless the postconfig
"microshift" is enabled). Hence, letting arguments.py resolve the
default value and checking that the file exists, is not best.

Instead, check for the file in ClusterDeployer.system_check(). This is
the place for such checks, and it can only perform the check, if the
file is actually needed.

This way, the validation is also happening during

  python ./clustersConfig.py "$CLUSTER_FILE" --system-check