isoCluster,extraConfigDpu: remove all IPU specific code paths for deploying "marvell-dpu"
For now, shortcut all IPU specific code paths when deploying for a
Marvell DPU (both the cluster-kind "iso" and the postconfig
"dpu_operator_{host,dpu}").

This does not actually make the marvell VSP work, but it at least allows
the code to run through and partially install the DPU and the dpu-operator.

The next step will be to implement the Marvell specific parts (if any).
The following step will be trying to further unify any DPU specific
parts.