clusterHost: make ClusterHost.bridge optional
With cluster kind "iso", we don't have (and don't need) a VirBridge nor
a BridgeConfig.

Make it optional, so that we can create ClusterHost instances for those
cluster kinds.

Note that any access to the new property ClusterHost.bridge will blow up
with an exception. We obviously, only access the bridge property for the
cases where we need and have a VirBridge.

While at it, add ClusterHost.__slots__. The benefit is that mypy will
complain if we try to set unexpected fields.