common: don't have default "repo_dir" in common.git_repo_setup()
A default value of "/root/dpu-operator" is bad, because it is only right
for a particular git repository.

The parameter is an Optional[str], so a better default would be None and
let the function autodetect the value from the URL. However, then we
would also need to return the detected path to the caller. Instead, make
the parameter required and not having a default value and not an
Optional[].