clustersConfig: add NodeConfig.dpu_kind for selecting type of DPU
When we install a DPU, we select cluster-kind "iso". Previously, we then
had the master node either as node-kind "physical" (for Intel IPU) or
"marvell-dpu".

At least in this case, CDA needed to know which kind of DPU we are
installing. This probably should not be detected, so we configured it in
the cluster configuration.

Later, with "postconfig: dpu_operator_dpu", we install the DPU operator
on the DPU. That works with the same cluster-kind "iso" and node-kind
"physical"/"marvell-ipu". In the future, we probably want that
installing the dpu-operator no longer requires CDA to know the DPU kind.
For now, it still does (and it can know it, since we already need to
know it to redeploy the DPU).

However, with "postconfig: dpu_operator_host", we currently need to know
which DPU we are installing on the host. As said, in the future,
dpu-operator may internally handle this, and CDA does not need to know.
For now, it needs to know.

Take this opportunity to clean this up by introducing a "dpu_kind".

If "dpu_kind" is unset, then it's as before. Cluster-kind "iso" and the
post configurations "dpu_operator_{dpu,host}" require the "dpu_kind".
The "dpu_kind" tells whether this is a Marvell/Intel DPU and whether
it's the host or the ACC. In those cases, the node-kind will be always
"physical".