isoCluster: handle iso file with Marvell DPU "iso" cluster
The podman container from "marvell-octeon-10-tools" supports filenames,
HTTP/HTTPS and "rhel:" values to specify the ISO file. For filenamees,
they need to be on the host where the container is running.

If CDA has a local file as iso file, we need to bring it over somehow.
We could start a HTTP server to share the file. However, that seems more
cumbersome, instead just copy the file over and let it use the local
path name.