clusterInfo: print cluster info in JSON
It's useful to see the loaded cluster info in the debug log.

Print it in JSON, so it is structured and we can easier inspect
it. Previously, not much was printed at all, because the ClusterInfo
class didn't implement a to-string method. Now, make it a dataclass,
which can be converted to JSON.