logger: use helpers from ktoolbox.common in logger module
ktoolbox.common has an ExtendedLogger class, similar to the one
we had here (and it is in fact taken from here with some adjustments).
Use it.

Also use log_config_logger() instead of configure_logger().

While the code is slightly different, there should be almost no change
in behavior. Changes:

- the "--help" output changes slightly.
- KTOOLBOX_ALL_LOGGERS=true is honored.
- with `python clustersConfig.py`, KTOOLBOX_LOGLEVEL is honored. Note
  that is not the case with `python cda.py`, where the argparser always
  explicitly selects a level one way or another.
- the logger for "ktoolbox" is now also enabled. However, since for
  now very few API from ktoobox is used (none at all?), this has little
  effect.