clustersConfig: support "{builder,base}_image" without TLS verification
One major point of making the builder/base image configurable is to be
able to run a local container registry.

In that case, it is probably self signed, and we would have to install
the certificates, so that `podman pull` works.

I wouldn't want on the system to place testing certificates into /etc to
be globally trusted and thereby degrading security for all of TLS. It
also seems too cumbersome to support configuring a separate cert-dir
(maybe one day). Instead, most likely I use CDA for testing and don't
actually care so much about security for the cluster (I still care for
security on my system otherwise).

Hence, add a way to pass "--tls-verify=false" to podman. Namely, if the
images are prefixed with a "!".