marvell: honor CDA_MARVELL_TOOLS_EXTRA_ARGS variable for pxeboot
We will soon want to test a branch where the Marvell VSP runs the
octep-cp-agent. In that case, we will want to pass
"--octep-cp-agent-service-disable" to pxeboot command.

Eventually, we will change the default behavior of pxeboot command, so
that "--octep-cp-agent-service-disable" is again done by default.

But there is a transiction period, where some runs will want to pass
this option, and others not.

By honoring the CDA_MARVELL_TOOLS_EXTRA_ARGS variable, the caller can
select the behavior we want.