marvell: redirect output of pxeboot also to /tmp file
cluster-deployment-automation by default runs with only INFO logging
level. Also, the output of `rsh.run()` is not easy to read.
In addition, redirect
the output of the command also to "/tmp/pxeboot-log-*" file, so that we
can look it up there too.

Granted, after reboot this log is gone. And also, with `make e2e-test`
we usually will afterwards reinstall the worker node and reboot the
machine. Thus, usually this log is soon lost after. It can still be
useful to be able to find the output in certain cases.